+++ internal/rules/rules.go
@@ -2,7 +2,72 @@
 // rule types together with their rule evaluation logic.
 package rules
 
-import "github.com/hashicorp/go-set/v3"
+// Meta describes metadata common to all available rules.
+type Meta struct {
+	// Description may contain the description of the rule. If the description
+	// is not empty, it is used to enrich error messages presented to the user.
+	Description string
+	// URL may contain a link to some (internal) documentation that further
+	// explains the purpose of a rule. If non-empty, it is used to enrich error
+	// messages presented to the user.
+	URL string
+}
+
+// Kind represents the kind of a rule.
+type Kind string
+
+const (
+	// KindAllOf asserts that all of the nested rules match.
+	KindAllOf Kind = "allOf"
+	// AnyOf asserts that at least one of the nested rules matches.
+	KindAnyOf Kind = "anyOf"
+	// Match defines a string to match trust anchors against.
+	KindMatch Kind = "match"
+	// MatchRegex defines a regular expression to match trust anchors against.
+	KindMatchRegex Kind = "matchRegex"
+	// Not inverts the matching behaviour of a rule.
+	KindNot Kind = "not"
+	// OneOf asserts that exactly one of the nested rules matches.
+	KindOneOf Kind = "oneOf"
+)
+
+// Rule is the interface implemented by all available rules.
+type Rule interface {
+	EvalRule
+	MetaRule
+	// Kind returns the kind of the rule.
+	Kind() Kind
+}
+
+// EvalRule is a rule that can be evaluated.
+type EvalRule interface {
+	// Eval evaluates the rule using the provided EvalContext.
+	Eval(ctx *EvalContext) EvalResult
+}
+
+// MetaRule provides setters and getters for rule metadata.
+type MetaRule interface {
+	// Meta returns the metadata associated with a rule.
+	Meta() Meta
+	// SetMeta sets the rule metadata.
+	SetMeta(meta Meta)
+}
+
+// metaRule is an implementation of MetaRule that's used by all available rules
+// to reduce boilerplate.
+type metaRule struct {
+	meta Meta
+}
+
+// Meta implements MetaRule.
+func (r *metaRule) Meta() Meta {
+	return r.meta
+}
+
+// SetMeta implements MetaRule.
+func (r *metaRule) SetMeta(meta Meta) {
+	r.meta = meta
+}
 
 // Ensure that all rule types implement the Rule interface.
 var (
@@ -13,38 +78,3 @@ var (
 	_ Rule = &NotRule{}
 	_ Rule = &OneOfRule{}
 )
-
-// emptyStringSet is a helper to create an empty string set. This is mainly
-// used to avoid verbose type hints at the call sites because set.From returns
-// a set.Set, but we actually work with the set.Collection interface.
-func emptyStringSet() set.Collection[string] {
-	return set.From([]string{})
-}
-
-// evalRulesResult is a helper type returned by evalRules.
-type evalRulesResult struct {
-	results      []EvalResult
-	matched      set.Collection[string]
-	successCount int
-}
-
-// evalRules evaluates a slice of rules and collects the results along with the
-// number of successes and a set of matched trust anchors.
-func evalRules(ctx *EvalContext, rules []Rule) evalRulesResult {
-	matched := emptyStringSet()
-	successCount := 0
-	results := make([]EvalResult, len(rules))
-
-	for i, rule := range rules {
-		result := rule.Eval(ctx)
-
-		if result.Success {
-			matched.InsertSet(result.Matched)
-			successCount++
-		}
-
-		results[i] = result
-	}
-
-	return evalRulesResult{results, matched, successCount}
-}
